Privilege Transfer
I'm adding the concept of "transfer of privilege". By that, I mean that allies have to be willing to transfer the benefits of their privilege (gender privilege, racial privilege, class privilege, etc.) to those without their privilege in order to create a more equitable and just system. I think this is a crucial concept for allies to understand.