Swap don't's and do's order to be positive-first
This is a suggested change to put the Do's of being a good ally first, and the Don't's second, so that people are first given a constructive, helpful and positive list, before being told what things not to do.  It does not alter the contents of either list. (And now, also, it matches the sentence introducing the two lists: "Here are some do's and don'ts…")