Encourage people to not lose sight of the purpose.
To get to the root of all evil can come from asking "why". We're focusing so much on the "what" with all the ruckus the "why" is getting buried.

"Why" are people looting? Maybe it's because we made them so poor it's come to that: https://www.youtube.com/watch?v=e68CoE70Mk8.