replace the bit about "remove politics"
hello - i noticed in the section on apologies that there was a mention of removing politics. i think i understand the intent here, to let go of politics for a moment when thinking about why we apologize. but the idea of removing politics bothers me for a lot of reasons - especially when considering recent events with Basecamp declaring their business as a no-politics area, and the massive problems this creates. 

i'd love to see the wording about removing politics removed from the guide and replaced with something less ambiguous. my suggestion in this PR is an idea that i had on what the wording may sound like, but i'm definitely open to any and all feedback around this and changes to the wording.